+++ packages/common/qds-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qualcomm-ui/qds-core Changelog
 
+## 1.27.2
+
+May 1st, 2026
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core, @qualcomm-ui/utils]
+
 ## 1.27.1
 
 May 1st, 2026